backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
python-dotenv==1.0.1
supabase>=2.12.0
neo4j==5.26.0
ultralytics==8.3.56
opencv-python-headless==4.10.0.84
Pillow==11.0.0
numpy==1.26.4
google-genai>=1.0.0
azure-cognitiveservices-speech>=1.38.0
